Consolidate config field listing code
All the directives, and the node formatters that they use, are now in a
common configfieldlists module. This is more in keeping with the
organization of the other directives in the lssttasks sub-package.
